Splitting the library into modules
